feat(#10): Perf Optimiztion
* I tried to optimize React excess rendering with help of useSelect and
reselect library, but I did not succeed. I have a suspicion that it's
somehow related to Object.values() that I use in reselect. But there's
antoher possibility. And it's probably related to using pagination in
the app. OR it might be both of them😁. Anyway I will try to create
another app with an SQL database. The reason I'm using Object.values()
is that i get an object of objects from Firebase and to render it in
React I need to transform it into an array. Naybe if I move
Object.values() from a selector into a component logic and simply pass
that container object into the component it might do the trick🤔